# Conflicts:
#	src/main/java/com/example/supergame/config/ItemsDatabaseConfig.java
#	src/main/java/com/example/supergame/config/PlayerDatabaseConfig.java
#	src/main/java/com/example/supergame/controller/ShopController.java
#	src/main/java/com/example/supergame/model/WeaponFactory.java
#	src/main/java/com/example/supergame/model/dto/item/MeleeWeapon.java
#	src/main/java/com/example/supergame/model/dto/item/RangeWeapon.java
#	src/main/java/com/example/supergame/model/dto/item/Weapon.java
#	src/main/java/com/example/supergame/model/enums/WeaponCategory.java
#	src/main/java/com/example/supergame/model/enums/WeaponType.java
#	src/main/java/com/example/supergame/repository/WeaponRepository.java
#	src/main/java/com/example/supergame/service/PlayerService.java
#	src/main/java/com/example/supergame/service/ShopService.java
#	src/test/java/com/example/supergame/service/PlayerServiceTest.java
#	src/test/java/com/example/supergame/testdata/TestDataFactory.java

+++ src/main/java/com/example/supergame/model/dto/item/Consumables.java
@@ -7,7 +7,7 @@ import org.springframework.data.mongodb.core.mapping.Document;
 public abstract class Consumables extends Item {
 
 
-    Consumables(String name, double price) {
+    Consumables(String name, int price) {
         super(name, price);
     }
 }
