+++ src/main/java/com/example/supergame/model/dto/item/MeleeWeapon.java
@@ -11,7 +11,7 @@ public class MeleeWeapon extends Weapon {
 
     private double damageMultiplier;
 
-    public MeleeWeapon(String name, double price, double damage, double accuracy, WeaponCategory weaponCategory, Rarity rarity, double damageMultiplier) {
+    public MeleeWeapon(String name, int price, double damage, double accuracy, WeaponCategory weaponCategory, Rarity rarity, double damageMultiplier) {
         super(name, price, damage, accuracy, weaponCategory, rarity);
         this.damageMultiplier = damageMultiplier;
     }
